pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2md"
version = "1.0.0"
description = "Convert PDF files to Markdown using the Mistral OCR API."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "pdf2md contributors"},
]
keywords = ["pdf", "markdown", "ocr", "mistral", "converter"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: General",
    "Topic :: Utilities",
]
dependencies = [
    "mistralai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pdf2md = "pdf2md.cli:main"

[project.urls]
Homepage = "https://github.com/your-username/pdf2md"
Repository = "https://github.com/your-username/pdf2md"
Issues = "https://github.com/your-username/pdf2md/issues"

[tool.setuptools.packages.find]
include = ["pdf2md*"]
